OK - the BasicArray class is done.  Get and set methods work and are quite efficient.  Next step is to complete the Variant class layer on top of it.
git-svn-id: https://svn.code.sf.net/p/freemat/code/branches/array-rework@3173 58e251e6-400e-0410-b350-93cf1ec51ee9

+++ FreeMat2/libs/libFreeMat/demo/Global.hpp
@@ -2,9 +2,10 @@
 #define __Global_hpp__
 
 #include <QtGlobal>
+#include "Types.hpp"
 
-//typedef double index_t;
-typedef int index_t;
+typedef double index_t;
+//typedef int64 index_t;
 const int NDims = 6;
 typedef QString Exception;
